requirements.txt
asgiref==3.5.2
Django==3.2.13
django-split-settings==1.2.0
flake8==6.0.0
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
